Fix like  logic and display posts
