+++ src/containers/Posts/Index.jsx
@@ -3,25 +3,9 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import _ from 'lodash';
 import moment from 'moment';
-import CircularProgress from 'material-ui/CircularProgress';
 import { getAllPosts, likePost } from '../../actions/post';
-import SearchPostPreview from '../../components/Post/SearchPostPreview';
+import DisplayPosts from '../../components/Post/DisplayPosts';
 import CreatePostButton from '../../components/CreatePostButton';
-
-const styles = {
-  postsContainer: {
-    display: 'flex',
-    justifyContent: 'center',
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    width: '50%',
-  },
-  postsWrapper: {
-    display: 'flex',
-    justifyContent: 'center',
-    flexDirection: 'row',
-  },
-};
 
 const propTypes = {
   isLoading: PropTypes.bool,
@@ -40,45 +24,15 @@ class AllPosts extends Component {
   }
 
   render() {
-    const { posts, isLoading } = this.props;
+    const { isLoading, userId, posts, actions } = this.props;
     return (
-      <div style={styles.postsWrapper}>
-        {
-          isLoading ? (
-            <div className="spinner">
-              <CircularProgress size={60} thickness={6} />
-            </div>
-          ) : ''
-        }
-        {
-          (!isLoading && _.isEmpty(posts)) ?
-          (<div className="no-posts-text">Dont have posts</div>) : ' '
-        }
-        {
-          (!isLoading && !_.isEmpty(posts)) ? (
-            <div style={styles.postsContainer}>
-              {
-                posts.map(post => (
-                  <SearchPostPreview
-                    key={post.id}
-                    user={post.author.name}
-                    title={post.title}
-                    description={post.description}
-                    avatar={post.author.image.data}
-                    time={post.time}
-                    likes={post.likes}
-                    comments={post.comments}
-                    createdDate={post.created_date}
-                    likePost={this.props.actions.likePost}
-                    userId={this.props.userId}
-                    postId={post.id}
-                    isLiked={post.isLiked}
-                  />
-                ))
-              }
-            </div>
-          ) : ' '
-        }
+      <div>
+        <DisplayPosts
+          isLoading={isLoading}
+          userId={userId}
+          posts={posts}
+          likePost={actions.likePost}
+        />
         <div>
           <CreatePostButton />
         </div>
