+++ src/containers/Posts/UserPosts.jsx
@@ -3,31 +3,9 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import _ from 'lodash';
 import moment from 'moment';
-import CircularProgress from 'material-ui/CircularProgress';
-import { getPostsByAuthor } from '../../actions/post';
-import SearchPostPreview from '../../components/Post/SearchPostPreview';
+import { getPostsByAuthor, likePost } from '../../actions/post';
+import DisplayPosts from '../../components/Post/DisplayPosts';
 import CreatePostButton from '../../components/CreatePostButton';
-
-const styles = {
-  noneText: {
-    margin: 'auto',
-    fontWeight: 100,
-    fontSize: 50,
-    color: '#cccccc',
-  },
-  postsContainer: {
-    display: 'flex',
-    justifyContent: 'center',
-    flexDirection: 'row',
-    flexWrap: 'wrap',
-    width: '50%',
-  },
-  postsWrapper: {
-    display: 'flex',
-    justifyContent: 'center',
-    flexDirection: 'row',
-  },
-};
 
 const propTypes = {
   isLoading: PropTypes.bool,
@@ -35,6 +13,7 @@ const propTypes = {
   posts: PropTypes.array,
   actions: PropTypes.shape({
     getPostsByAuthor: PropTypes.func,
+    likePost: PropTypes.func, // eslint-disable-line
   }),
 };
 
@@ -45,41 +24,15 @@ class UserPosts extends Component {
   }
 
   render() {
-    const { posts, isLoading } = this.props;
+    const { isLoading, userId, posts, actions } = this.props;
     return (
-      <div style={styles.postsWrapper}>
-        {
-          isLoading ? (
-            <div className="spinner">
-              <CircularProgress size={60} thickness={6} />
-            </div>
-          ) : ''
-        }
-        {
-          (!isLoading && _.isEmpty(posts)) ?
-          (<div className="no-posts-text">Dont have posts</div>) : ' '
-        }
-        {
-          (!isLoading && !_.isEmpty(posts)) ? (
-            <div style={styles.postsContainer}>
-              {
-                posts.map(post => (
-                  <SearchPostPreview
-                    key={post.id}
-                    user={post.author.name}
-                    title={post.title}
-                    description={post.description}
-                    avatar={post.author.image.data}
-                    time={post.time}
-                    likes={post.likes}
-                    comments={post.comments}
-                    createdDate={post.created_date}
-                  />
-                ))
-              }
-            </div>
-          ) : ' '
-        }
+      <div>
+        <DisplayPosts
+          isLoading={isLoading}
+          userId={userId}
+          posts={posts}
+          likePost={actions.likePost}
+        />
         <div>
           <CreatePostButton />
         </div>
@@ -108,6 +61,7 @@ export default connect((state) => {
     },
     time: post.time,
     likes: post.likes.length,
+    isLiked: _.includes(post.likes, userId) ? true : false, // eslint-disable-line
     comments: 0,
     created_date: moment(post.created_date).format('ll'),
   })) : [];
@@ -120,5 +74,6 @@ export default connect((state) => {
 }, dispatch => ({
   actions: bindActionCreators({
     getPostsByAuthor,
+    likePost,
   }, dispatch),
 }))(UserPosts);
